requirement.txt
absl-py==1.0.0
appdirs==1.4.4
audioread==2.1.9
backcall==0.2.0
cachetools==4.2.4
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
ffmpeg==1.4
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grpcio==1.42.0
idna==3.2
imageio==2.9.0
imageio-ffmpeg==0.4.5
importlib-metadata==4.8.2
imutils==0.5.4
ipython==7.28.0
jedi==0.18.0
joblib==1.0.1
kiwisolver==1.3.2
librosa==0.8.1
llvmlite==0.37.0
Markdown==3.3.6
matplotlib==3.4.3
matplotlib-inline==0.1.3
moviepy==1.0.3
numba==0.54.0
numpy==1.20.3
oauthlib==3.1.1
opencv-contrib-python==4.5.4.58
opencv-python==4.5.3.56
packaging==21.0
pandas==1.1.5
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.2
pooch==1.5.1
proglog==0.1.9
prompt-toolkit==3.0.20
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.10.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytube==11.0.1
pytz==2021.1
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.7.2
scikit-learn==1.0
scipy==1.1.0
seaborn==0.11.2
six==1.16.0
SoundFile==0.10.3.post1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
thop==0.0.31.post2005241907
threadpoolctl==2.2.0
torch==1.10.0
torchvision==0.11.1
tqdm==4.62.3
traitlets==5.1.0
typing_extensions==4.0.0
urllib3==1.26.7
wcwidth==0.2.5
Werkzeug==2.0.2
yapf==0.31.0
you-get==0.4.1545
zipp==3.6.0
